+++ fork/engine/MagicSqInfo.java
@@ -3,5 +3,5 @@ package fork.engine;
 public class MagicSqInfo {
     long magicNo;
     long blockerMask;
-    byte shift;
+    int shift;
 }
